+++ src/main/java/com/example/interntest_ma/entity/Receipt.java
@@ -1,6 +1,7 @@
 package com.example.interntest_ma.entity;
 
 import lombok.AccessLevel;
+import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
@@ -8,13 +9,15 @@ import lombok.experimental.FieldDefaults;
 import javax.persistence.Entity;
 import javax.persistence.Table;
 
+
 @Entity
 @Getter
 @Setter
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Table(name = "receipt")
 public class Receipt extends Auditing {
-    Long goodRices;
+    Float goodRices;
+
 
 
 }
